requirements.txt
flask
anthropic
requests
python-dotenv
gunicorn